requirements.txt
numpy
matplotlib
imageio
pygame-ce
